WIP: inverted index for time-series #54
